+++ fpraktikum/admin.py
@@ -93,7 +93,6 @@ class WaitlistResource(resources.ModelResource):
         fields = tuple(map(lambda x: x.name, model._meta.get_fields()))
 
 
-
 class FpWaitlistAdmin(admin.ModelAdmin):
     list_display = [f.name for f in FpWaitlist._meta.get_fields()]
