+++ fpraktikum/db_utils.py
@@ -6,7 +6,6 @@ import datetime
 import csv
 from django.http import HttpResponse
 from rest_framework import generics
-from import_export.views import ExportViewMixin
 
 def il_db_retrieve(user_lastname, user_login):
     """
@@ -46,24 +45,3 @@ def is_user_valid(login):
             return False
 
     return True
-
-
-class ExportView(generics.GenericAPIView, ExportViewMixin):
-    queryset = None
-    serializer_class = None
-    permission_classes = ()
-    resource_class = None
-    file_name = ""
-
-    def get(self, request, *args, **kwargs):
-        resource = self.resource_class()
-        data = resource.export()
-        headers = resource.get_export_headers()
-        response = HttpResponse(content_type="text/csv")
-        date = datetime.datetime.now().strftime("%d_%m_%Y")
-        response['Content-Disposition'] = 'attachment; filename={0}{1}.csv'.format(self.file_name, date)
-        writer = csv.writer(response, dialect=csv.excel)
-        writer.writerow(headers)
-        for registrant in data:
-            writer.writerow(registrant)
-        return response
