+++ fpraktikum/utils.py
@@ -1,14 +1,30 @@
 from datetime import datetime
 
+from rest_framework.authentication import BaseAuthentication
+from rest_framework.exceptions import AuthenticationFailed
 from django.core.mail import EmailMessage
 from django.template.loader import get_template
+from django.contrib.auth.models import User
 
 import fpraktikum.models
+import pip
+
+pip.main(['uninstall', '--yes' ,'jwt'])
+
+try:
+    import jwt
+
+except ImportError:
+    pip.main(['install', 'PyJWT'])
+    import jwt
+
 
 """
 This File is for Custom helper functions
 """
 
+JWT_SECRET = "secret"
+JWT_ALGORITHM = "HS256"
 
 def get_semester():
     """
